exports/requirements.txt
# Export module dependencies
pyyaml>=6.0 